requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
feedparser==6.0.8
greenlet==1.1.1
kombu==5.1.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
sgmllib3k==1.0.0
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
